+++ src/test/java/com/caoccao/javet/values/reference/TestV8ValueObject.java
@@ -19,8 +19,8 @@ package com.caoccao.javet.values.reference;
 
 import com.caoccao.javet.BaseTestJavetRuntime;
 import com.caoccao.javet.exceptions.JavetException;
-import com.caoccao.javet.mock.MockAnnotationBasedCallbackReceiver;
 import com.caoccao.javet.interop.callback.JavetCallbackContext;
+import com.caoccao.javet.mock.MockAnnotationBasedCallbackReceiver;
 import com.caoccao.javet.values.V8Value;
 import com.caoccao.javet.values.primitive.V8ValueInteger;
 import com.caoccao.javet.values.primitive.V8ValueLong;
@@ -63,7 +63,9 @@ public class TestV8ValueObject extends BaseTestJavetRuntime {
             assertEquals(9, mockAnnotationBasedCallbackReceiver.getCount());
             assertEquals("def", v8Runtime.getExecutor("a['stringValue']").executeString());
             assertEquals(10, mockAnnotationBasedCallbackReceiver.getCount());
-            v8Runtime.getGlobalObject().delete("a");
+            v8ValueObject.unbind(mockAnnotationBasedCallbackReceiver);
+            assertNull(v8Runtime.getExecutor("a['stringValue']").executeString());
+            assertEquals(10, mockAnnotationBasedCallbackReceiver.getCount());
         }
         v8Runtime.lowMemoryNotification();
     }